app/requirements/base.txt
autopep8==1.5.4
boto3==1.16.23
botocore==1.19.23
celery==4.4.6
dj-database-url==0.5.0
Django==3.1.5
django-cors-headers==3.5.0
django-filter==2.4.0
django-storages==1.10.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
docutils==0.16
flower==0.9.4
Markdown==3.3.3
MarkupSafe==1.1.1
packaging==20.4
psycopg2-binary==2.8.6
Pygments==2.7.2
PyJWT==1.7.1
python-dateutil==2.8.1
python-decouple==3.3
redis==3.5.3
s3transfer==0.3.3
sendgrid==6.4.7
whitenoise==5.2.0
email-validator==1.1.2
channels==3.0.3
channels-redis==3.2.0
pytest-asyncio==0.14.0
pytest-django==4.1.0
django-redis==4.12.1
locust==1.2.2
pylint==2.6.0
pyfcm==1.4.7
pyotp==2.4.1
googlemaps==4.4.2
sentry-sdk==0.19.4
django-import-export==2.4.0
twilio==6.48.0
drf-spectacular==0.11.1
Faker==4.17.1
selenium==3.141.0
django-graphql-jwt==0.3.1 
graphene==2.1.8 
graphene-django==2.15.0
django-graphene-permissions==0.0.4
graphene-file-upload==1.2.2
graphene-django-extras==0.4.9
django-graphql-social-auth==0.1.4